+++ src/components.d.ts
@@ -12,18 +12,10 @@ declare module 'vue' {
     AdminSectionCard: typeof import('./components/admin/AdminSectionCard.vue')['default']
     AdminToolBar: typeof import('./components/admin/AdminToolBar.vue')['default']
     AppToolBar: typeof import('./components/common/AppToolBar.vue')['default']
-    ChatHeader: typeof import('./components/chat/ChatHeader.vue')['default']
-    ChatInput: typeof import('./components/chat/ChatInput.vue')['default']
-    ChatNoteDialog: typeof import('./components/chat/ChatNoteDialog.vue')['default']
-    ChatNoteList: typeof import('./components/chat/ChatNoteList.vue')['default']
-    ChatSearchDialog: typeof import('./components/chat/ChatSearchDialog.vue')['default']
-    ChatShareDialog: typeof import('./components/chat/ChatShareDialog.vue')['default']
     ComplainDialog: typeof import('./components/chat/ComplainDialog.vue')['default']
     DetailView: typeof import('./components/common/DetailView.vue')['default']
-    MyChatMessage: typeof import('./components/chat/MyChatMessage.vue')['default']
     Popoto: typeof import('./components/common/Popoto.vue')['default']
     ProgressCard: typeof import('./components/dashboard/ProgressCard.vue')['default']
-    RecommendList: typeof import('./components/chat/RecommendList.vue')['default']
     RouterLink: typeof import('vue-router')['RouterLink']
     RouterView: typeof import('vue-router')['RouterView']
     SessionList: typeof import('./components/chat/SessionList.vue')['default']
