+++ src/layouts/MainLayout.vue
@@ -27,8 +27,6 @@
 </template>
 
 <script setup lang="ts">
-import MyChatMessageVue from "@/components/chat/MyChatMessage.vue";
-import MyChatMessage from "@/components/chat/MyChatMessage.vue";
 import SessionList from "@/components/chat/SessionList.vue";
 import AppToolBar from "@/components/common/AppToolBar.vue";
 import UIConfig from "@/components/common/UIConfig.vue";
