+++ src/views/chat/components/ChatNoteDialog.vue
