+++ src/views/chat/components/ChatShareDialog.vue
